server/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
supabase>=2.4.0
python-dotenv>=1.0.0
pydantic>=2.6.0
